split fee adjustment from 'CoinSelection' into a separate module 'CoinSelection.Fee'
